Refactor: ContextAPI 코드 정리
